refactor: break up SLO page template into smaller (reusable) template parts
